流量削峰： 1. 引入秒杀令牌，秒杀活动模块负责秒杀令牌的生成，用户风控策略以及库存售罄判断前置到秒杀令牌发放中 2. 引入秒杀大闸，根据秒杀商品初始库存颁发对应数量令牌，控制大闸流量 3. 本地队列泄洪

+++ src/main/java/com/miaoshaproject/mq/MqProducer.java
@@ -92,9 +92,9 @@ public class MqProducer {
                 if (Objects.isNull(stockLogDO)) {
                     return LocalTransactionState.UNKNOW;
                 }
-                if (stockLogDO.getStatus().intValue() == 2) {
+                if (stockLogDO.getStatus() == 2) {
                     return LocalTransactionState.COMMIT_MESSAGE;
-                } else if (stockLogDO.getStatus().intValue() == 1) {
+                } else if (stockLogDO.getStatus() == 1) {
                     return LocalTransactionState.UNKNOW;
                 }
                 return LocalTransactionState.ROLLBACK_MESSAGE;
